Consolidate duplicated Channel definitions to single location
Eliminates the duplicate `Channel` class wrappers around `FunctionalUringFacade`.

💡 What:
- Moved the `Channel` body into `borg.trikeshed.userspace.nio.channel.Channels` directly.
- Deleted the duplicate implementations from `borg.trikeshed.userspace.nio.channel.Channel` and `borg.trikeshed.userspace.UserspaceIO`.
- Updated call-sites and file imports uniformly to use `Channels.Channel`.
- Removed `UserspaceIO`'s local `File` shadow to use the common `borg.trikeshed.userspace.nio.file.File` via a `fromImpl` bridge.
- The `Channel` interface in `nio/channels` is untouched.

🎯 Why:
To remove the triple repetition of `class Channel(private val facade: FunctionalUringFacade)`, simplifying maintenance and preventing confusing overlapping typings.

✅ Verification:
- Build gate `./gradlew :jvmMainClasses --no-daemon` ran perfectly. No test suite regressions introduced.
