+++ src/commonMain/kotlin/borg/trikeshed/lib/SeekFileBufferCommon.kt
@@ -2,8 +2,8 @@ package borg.trikeshed.lib
 
 import borg.trikeshed.lib.long.LongSeries
 import borg.trikeshed.userspace.ByteRegion
-import borg.trikeshed.userspace.nio.channel.Channel
-import borg.trikeshed.userspace.nio.channel.Channels
+import borg.trikeshed.userspace.Channel
+import borg.trikeshed.userspace.Channels
 import borg.trikeshed.userspace.nio.file.File
 import borg.trikeshed.userspace.nio.file.Files
 import borg.trikeshed.userspace.nio.ByteBuffer
