+++ src/commonMain/kotlin/borg/trikeshed/userspace/UserspaceIO.kt
@@ -14,39 +14,6 @@ data class SelectionResult(val res: Int, val userData: Long)
  * The typed API is sugar that creates [UringSubmission] internally.
  * New code should use the unified path exclusively.
  */
-class Channel(
-    private val facade: FunctionalUringFacade,
-) {
-    fun read(file: File, buffer: ByteBuffer, offset: Long, userData: Long) =
-        facade.read(file.impl, buffer, offset, userData)
-
-    fun write(file: File, buffer: ByteBuffer, offset: Long, userData: Long) =
-        facade.write(file.impl, buffer, offset, userData)
-
-    fun accept(file: File, userData: Long) =
-        facade.accept(file.impl, userData)
-
-    fun connect(file: File, address: String, port: Int, userData: Long) =
-        facade.connect(file.impl, address, port, userData)
-
-    fun close(file: File, userData: Long) =
-        facade.close(file.impl, userData)
-
-    fun sync(file: File, userData: Long, metaData: Boolean) =
-        facade.sync(file.impl, userData, metaData)
-
-    fun truncate(file: File, size: Long, userData: Long) =
-        facade.truncate(file.impl, size, userData)
-
-    fun map(file: File, mode: String, position: Long, size: Long, userData: Long) =
-        facade.map(file.impl, mode, position, size, userData)
-
-    fun submit(): Int = facade.submit()
-
-    fun wait(minComplete: Int = 1): List<SelectionResult> = facade.wait(minComplete)
-
-    fun peek(): List<SelectionResult> = facade.peek()
-}
 
 class File internal constructor(internal val impl: FileImpl) {
     val id: Int get() = impl.id
