+++ src/commonMain/kotlin/borg/trikeshed/userspace/nio/channels/DatagramChannel.kt
@@ -4,8 +4,8 @@ package borg.trikeshed.userspace.nio.channels
 
 import borg.trikeshed.userspace.ByteRegion
 import borg.trikeshed.userspace.nio.ByteBuffer
-import borg.trikeshed.userspace.nio.channel.Channel
-import borg.trikeshed.userspace.nio.channel.Channels
+import borg.trikeshed.userspace.Channel
+import borg.trikeshed.userspace.Channels
 import borg.trikeshed.userspace.nio.channels.spi.AbstractSelectableChannel
 import borg.trikeshed.userspace.nio.channels.spi.SelectorProvider
 import borg.trikeshed.userspace.nio.file.File
