+++ src/commonMain/kotlin/borg/trikeshed/userspace/nio/channels/SocketChannel.kt
@@ -2,8 +2,8 @@
 
 package borg.trikeshed.userspace.nio.channels
 
-import borg.trikeshed.userspace.nio.channel.Channel
-import borg.trikeshed.userspace.nio.channel.Channels
+import borg.trikeshed.userspace.Channel
+import borg.trikeshed.userspace.Channels
 import borg.trikeshed.userspace.nio.file.File
 import borg.trikeshed.userspace.nio.ByteBuffer
 import borg.trikeshed.userspace.nio.channels.spi.AbstractSelectableChannel
