+++ src/commonMain/kotlin/borg/trikeshed/userspace/nio/channels/UringFileChannel.kt
@@ -1,6 +1,6 @@
 package borg.trikeshed.userspace.nio.channels
 
-import borg.trikeshed.userspace.nio.channel.Channel
+import borg.trikeshed.userspace.Channel
 import borg.trikeshed.userspace.nio.file.File
 import borg.trikeshed.userspace.nio.ByteBuffer
