+++ src/commonMain/kotlin/borg/trikeshed/userspace/nio/file/File.kt
@@ -5,12 +5,13 @@ import borg.trikeshed.userspace.FileImpl
 /**
  * Open file handle — backed by [FileImpl] (expect/actual).
  *
- * Used by [borg.trikeshed.userspace.nio.channel.Channel] operations.
+ * Used by [borg.trikeshed.userspace.Channel] operations.
  * For directory-level operations use [Files].
  */
 class File internal constructor(internal val impl: FileImpl) {
     companion object {
         internal fun fromFd(fd: Int): File = File(FileImpl(fd))
+        fun fromImpl(impl: FileImpl): File = File(impl)
     }
 
     val id: Int get() = impl.id
